test module added

+++ modules/common/err_datum.py
@@ -21,13 +21,9 @@ def err_datum_path(err: str,DirDatm: Path,DirErrBase: Path,RmvDatmOut: bool,DirO
     :        the #/pfs/BASE_REPO portion of DirDatm with DirErrBase.
     """
 
-    log.debug(err)
     caller = inspect.stack()[1].function
-    log.info(f'Error resulted from call: {caller}')
-
-    # Inform the user of the error routing
-    log.info(f'Re-routing failed datum .........')
-    log.info(f'      from: {DirDatm}')
+    log.info(f'The error, "{err}", resulted from call: {caller}. ')
+    log.info(f'Rerouting starts..... ')
 
     # call to parse input directory
     DirInInfo = get_dir_info(DirIn = DirDatm)
@@ -37,19 +33,28 @@ def err_datum_path(err: str,DirDatm: Path,DirErrBase: Path,RmvDatmOut: bool,DirO
     if DirOutBase == None:
         DirOutBase = Path(DirErrBase).parents[0]
     #
-    DirRepo = DirInInfo[3]
-    DirErr_path = Path(DirErrBase,DirRepo)
-    DirOut_path = Path(DirOutBase,DirRepo)
-    os.makedirs(DirErr_path,exist_ok=True)
-    Err_file = os.path.join(DirErr_path,os.path.basename(DirErr_path).split('/')[-1])
+    if DirInInfo == []:
+        log.info(f'Re-routing stopped due to the input path structure not compliant.........\n')
+
+    else:
+        # Inform the user of the error routing
+        log.info(f'Re-routing failed datum .........')
+        log.info(f'      from: {DirDatm}')
+        DirRepo = DirInInfo[3]
+        DirErr_path = Path(DirErrBase,DirRepo)
+        log.info(f'      to: {DirErr_path}')
+        DirOut_path = Path(DirOutBase,DirRepo)
+        os.makedirs(DirErr_path,exist_ok=True)
+        Err_file = os.path.join(DirErr_path,os.path.basename(DirErr_path).split('/')[-1])
     # Write an empty file
-    file1 = open(Err_file,"w")
-    file1.close()
-    log.info(f'An empty file is written in:  {DirErr_path}')
+        file1 = open(Err_file,"w")
+        file1.close()
+        log.info(f'An empty file is written in:  {DirErr_path}')
+        log.info(f'Rerouting completed successfully..... \n')
     #
     # Remove any partial output for the datum
-    if (RmvDatmOut == True):
-        if os.path.exists(DirOut_path):
-            removed = shutil.rmtree(DirOut_path)
-            log.info(f'Removed partial output for errored datum:  {DirOut_path}')
-            print('\n\t ====== removed: ',removed)
+        if (RmvDatmOut == True):
+            if os.path.exists(DirOut_path):
+                removed = shutil.rmtree(DirOut_path)
+                log.info(f'Removed partial output for errored datum:  {DirOut_path}')
+                print('\n\t ====== removed: ',removed)
